requirements.txt
ccxt==4.1.13
python-dotenv==1.0.0
toml==0.10.2 